@@ -32,86 +32,86 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.16.5",
-    "@babel/preset-env": "^7.16.5",
-    "@babel/preset-react": "^7.16.5",
-    "@testing-library/react": "^12.1.2",
+    "@babel/core": "^7.17.10",
+    "@babel/preset-env": "^7.17.10",
+    "@babel/preset-react": "^7.16.7",
+    "@testing-library/react": "^13.2.0",
     "@types/classnames": "^2.3.1",
-    "@types/history": "^4.7.9",
-    "@types/jest": "^27.0.3",
-    "@types/node": "^17.0.0",
-    "@types/react": "^17.0.37",
-    "@types/react-dom": "^17.0.11",
-    "@types/react-redux": "^7.1.20",
-    "@types/react-router": "^5.1.17",
+    "@types/history": "^5.0.0",
+    "@types/jest": "^27.5.0",
+    "@types/node": "^17.0.32",
+    "@types/react": "^18.0.9",
+    "@types/react-dom": "^18.0.3",
+    "@types/react-redux": "^7.1.24",
+    "@types/react-router": "^5.1.18",
     "@types/redux-actions": "^2.6.2",
     "@types/webpack": "^5.28.0",
-    "@types/webpack-env": "^1.16.3",
-    "@typescript-eslint/eslint-plugin": "^5.7.0",
-    "@typescript-eslint/parser": "^5.7.0",
-    "babel-jest": "^27.4.5",
-    "babel-loader": "^8.2.3",
+    "@types/webpack-env": "^1.16.4",
+    "@typescript-eslint/eslint-plugin": "^5.23.0",
+    "@typescript-eslint/parser": "^5.23.0",
+    "babel-jest": "^28.1.0",
+    "babel-loader": "^8.2.5",
     "clean-webpack-plugin": "^4.0.0",
-    "css-loader": "^6.5.1",
+    "css-loader": "^6.7.1",
     "enzyme": "^3.11.0",
-    "eslint": "^8.4.1",
-    "eslint-config-standard": "^16.0.3",
-    "eslint-plugin-import": "^2.25.3",
+    "eslint": "^8.15.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.2.0",
-    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.29.4",
     "file-loader": "^6.2.0",
-    "gh-pages": "^3.2.3",
-    "html-loader": "^3.0.1",
+    "gh-pages": "^4.0.0",
+    "html-loader": "^3.1.0",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^27.4.5",
-    "mini-css-extract-plugin": "^2.4.5",
-    "postcss": "^8.4.5",
+    "jest": "^28.1.0",
+    "mini-css-extract-plugin": "^2.6.0",
+    "postcss": "^8.4.13",
     "postcss-browser-reporter": "^0.6.0",
-    "postcss-import": "^14.0.2",
+    "postcss-import": "^14.1.0",
     "postcss-loader": "^6.2.1",
-    "postcss-preset-env": "^7.0.2",
-    "postcss-reporter": "^7.0.4",
+    "postcss-preset-env": "^7.5.0",
+    "postcss-reporter": "^7.0.5",
     "postcss-url": "^10.1.3",
-    "prettier": "^2.5.1",
+    "prettier": "^2.6.2",
     "react-hot-loader": "^4.13.0",
     "redux-devtools-extension": "^2.13.9",
     "style-loader": "^3.3.1",
-    "ts-loader": "^9.2.6",
-    "ts-node": "^10.4.0",
-    "typescript": "^4.5.4",
+    "ts-loader": "^9.3.0",
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.4",
     "url-loader": "^4.1.1",
-    "webpack": "^5.65.0",
-    "webpack-cli": "^4.9.1",
-    "webpack-dev-server": "^4.6.0",
-    "yarn-upgrade-all": "^0.5.4"
+    "webpack": "^5.72.1",
+    "webpack-cli": "^4.9.2",
+    "webpack-dev-server": "^4.9.0",
+    "yarn-upgrade-all": "^0.7.1"
   },
   "dependencies": {
-    "@babel/plugin-transform-runtime": "^7.16.5",
-    "@babel/preset-typescript": "^7.16.5",
-    "@types/lodash": "^4.14.178",
-    "@types/react-router-dom": "^5.3.2",
-    "axios": "^0.24.0",
+    "@babel/plugin-transform-runtime": "^7.17.10",
+    "@babel/preset-typescript": "^7.16.7",
+    "@types/lodash": "^4.14.182",
+    "@types/react-router-dom": "^5.3.3",
+    "axios": "^0.27.2",
     "classnames": "^2.3.1",
-    "copy-webpack-plugin": "^10.2.0",
-    "eslint-config-prettier": "^8.3.0",
-    "eslint-config-react-app": "^7.0.0",
+    "copy-webpack-plugin": "^10.2.4",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-react-app": "^7.0.1",
     "eslint-plugin-flowtype": "^8.0.3",
     "eslint-plugin-jsx-a11y": "^6.5.1",
     "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
     "lodash": "^4.17.21",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.6",
-    "react-router": "^6.1.1",
-    "react-router-dom": "^6.2.1",
-    "react-test-renderer": "^17.0.2",
-    "redux": "^4.1.2",
+    "react": "^18.1.0",
+    "react-dom": "^18.1.0",
+    "react-redux": "^8.0.1",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
+    "react-test-renderer": "^18.1.0",
+    "redux": "^4.2.0",
     "redux-actions": "^2.6.5",
     "redux-thunk": "^2.4.1",
-    "sass": "^1.45.0",
-    "sass-loader": "^12.4.0",
+    "sass": "^1.51.0",
+    "sass-loader": "^12.6.0",
     "whatwg-fetch": "^3.6.2"
   }
 }
