@@ -1,16 +1,18 @@
-import React from 'react'
-import ReactDOM from 'react-dom'
-import { Provider } from 'react-redux'
-import { configureStore } from 'app/store'
-import { AppRoutes } from './app/routes'
-import './app/styles/style.scss'
+import React from 'react';
+import { createRoot } from 'react-dom/client';
+import { Provider } from 'react-redux';
+import { configureStore } from 'app/store';
+import { AppRoutes } from './app/routes';
+import './app/styles/style.scss';
 
 // prepare store
-const store = configureStore()
+const store = configureStore();
 
-ReactDOM.render(
-  <Provider store={store}>
-    <AppRoutes />
-  </Provider>,
-  document.getElementById('root')
-)
+const root = createRoot(document.getElementById('root') as HTMLElement);
+root.render(
+  <React.StrictMode>
+    <Provider store={store}>
+      <AppRoutes />
+    </Provider>
+  </React.StrictMode>
+);
