@@ -4,16 +4,16 @@ import {
   Routes,
   Route
 } from 'react-router-dom'
-// import NotFound from './components/NotFound'
-import Timer from './components/Timer'
-import WeatherWidgetContainer from './components/WeatherWidget'
+import NotFound from './modules/NotFound'
+import Timer from './modules/Timer'
+import WeatherWidgetContainer from './modules/WeatherWidget'
 
 export const AppRoutes = () => (
   <BrowserRouter>
     <Routes>
       <Route path="/" element={<WeatherWidgetContainer />} />
       <Route path="/timer" element={<Timer />} />
-      {/* <Route path='*' element={<NotFound />} /> */}
+      <Route path='*' element={<NotFound />} />
     </Routes>
   </BrowserRouter>
 )
