@@ -2,6 +2,6 @@
 
 // for style loader
 declare module '*.scss' {
-  const styles: any;
+  const styles: any
   export = styles;
 }
